+++ ide/src/main/java/nl/xillio/migrationtool/XillioRuntimeException.java
@@ -1,0 +1,25 @@
+package nl.xillio.migrationtool;
+
+/**
+ * Custom Xillio runtime exception class
+ *
+ * Created by Anwar on 11/17/2015.
+ */
+public class XillioRuntimeException extends RuntimeException {
+
+    public XillioRuntimeException() {
+
+    }
+
+    public XillioRuntimeException(String message) {
+        super(message);
+    }
+
+    public XillioRuntimeException(String message, Throwable t) {
+        super(message, t);
+    }
+
+    public XillioRuntimeException(Throwable t) {
+        super(t);
+    }
+}