+++ ide/src/main/resources/ace/load-editor.js
@@ -34,7 +34,7 @@ function loadEditor(){
         var editor = contenttools.getAce();
         hl = editor.getSession().addMarker(new Range(line, 0, line, 1), "ace_" + type, "fullLine");
         editor.$highlights.push(hl);
-        editor.scrollToLine(line);
+        editor.scrollToLine(line, true);
     }
     editor.clearHighlight = function() {
         var editor = contenttools.getAce();
@@ -137,6 +137,7 @@ function loadEditor(){
 	// Remove keybindings
 	editor.commands.removeCommand('find');
 	editor.commands.removeCommand('replace');
+	editor.commands.removeCommand('removeline');
 	
 	editor.getSession().setUndoManager(new UndoManager());
 	editor.setOption("dragEnabled", false);
@@ -148,4 +149,11 @@ function loadEditor(){
 	}
 	editor.on('cut', toClipboard);
 	editor.on('copy', toClipboard);
+
+	// Duplicate the selected lines.
+	editor.duplicateCurrentLines = function() {
+		var range = editor.selection.getRange();
+		editor.session.duplicateLines(range.start.row, range.end.row);
+		editor.selection.setRange(range);
+	};
 }