+++ xill-plugins/plugin-file/src/main/java/nl/xillio/xill/plugins/file/constructs/IterateFilesConstruct.java
@@ -14,6 +14,9 @@ import nl.xillio.xill.plugins.file.utils.Folder;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.AccessDeniedException;
+import java.nio.file.NoSuchFileException;
+import java.nio.file.NotDirectoryException;
 import java.util.Iterator;
 
 /**
@@ -45,14 +48,13 @@ public class IterateFilesConstruct extends Construct {
 
 			return result;
 
-		} catch (IOException e) {
-			if ("java.nio.file.NoSuchFileException".equals(e.getClass().getName())) {
-				throw new RobotRuntimeException("The specified folder does not exist:  " + e.getMessage(), e);
-			} else if ("java.nio.file.AccessDeniedException".equals(e.getClass().getName())) {
-				throw new RobotRuntimeException("Access to the specified folder is denied:  " + e.getMessage(), e);
-			} else if ("java.nio.file.NotDirectoryException".equals(e.getClass().getName())) {
-				throw new RobotRuntimeException("The specified folder is not a directory:  " + e.getMessage(), e);
-			}
+		} catch (NoSuchFileException e) {
+            throw new RobotRuntimeException("The specified folder does not exist:  " + e.getMessage(), e);
+        } catch (AccessDeniedException e) {
+            throw new RobotRuntimeException("Access to the specified folder is denied:  " + e.getMessage(), e);
+        } catch (NotDirectoryException e) {
+            throw new RobotRuntimeException("The specified folder is not a directory:  " + e.getMessage(), e);
+        } catch (IOException e) {
 			throw new RobotRuntimeException("An error occurred: " + e.getMessage(), e);
 		}
 	}